2.5.0

+++ drivers/scsi/sym53c8xx_2/sym_glue.h
@@ -77,6 +77,7 @@
 #include <linux/errno.h>
 #include <linux/pci.h>
 #include <linux/string.h>
+#include <linux/slab.h>
 #include <linux/mm.h>
 #include <linux/ioport.h>
 #include <linux/time.h>
@@ -462,7 +463,7 @@ struct sym_shcb {
 
 	vm_offset_t	mmio_va;	/* MMIO kernel virtual address	*/
 	vm_offset_t	ram_va;		/* RAM  kernel virtual address	*/
-	u_long		io_port;	/* IO port address cookie	*/
+	u32		io_port;	/* IO port address		*/
 	u_short		io_ws;		/* IO window size		*/
 	int		irq;		/* IRQ number			*/
